bug fixes and documentation updates
