+++ cmd/plaxrun/dsl/test_run.go
@@ -176,7 +176,7 @@ func (tr *TestRun) Exec(ctx *Ctx) error {
 
 	testReport.Finish()
 
-	err = tr.Reports.Generate(ctx, testReport, *tr.trps.EmitJSON)
+	err = tr.Reports.Generate(ctx.Ctx, tr.Params, tr.trps.Bindings, testReport, *tr.trps.EmitJSON)
 	if err != nil {
 		ctx.Logf(err.Error())
 	}
